+++ hs-bindgen/fixtures/binding-specs/fun_arg/macro/function/Example/Unsafe.hs
@@ -17,13 +17,13 @@ $(HsBindgen.Runtime.Internal.CAPI.addCSource (HsBindgen.Runtime.Internal.CAPI.un
   , "  foo(arg1);"
   , "}"
   , "void hs_bindgen_cf67e2fc00fd28d8 ("
-  , "  A arg1"
+  , "  A *arg1"
   , ")"
   , "{"
   , "  fooA(arg1);"
   , "}"
   , "void hs_bindgen_269a46f9680e33ed ("
-  , "  B arg1"
+  , "  B *arg1"
   , ")"
   , "{"
   , "  fooB(arg1);"
@@ -76,12 +76,12 @@ foo = hs_bindgen_fbc2ec26cd297034
 
 -- __unique:__ @test_bindingspecsfun_argmacrofu_Example_Unsafe_fooA@
 foreign import ccall unsafe "hs_bindgen_cf67e2fc00fd28d8" hs_bindgen_cf67e2fc00fd28d8_base ::
-     (RIP.Int32 -> IO RIP.Int32)
+     RIP.FunPtr RIP.Void
   -> IO ()
 
 -- __unique:__ @test_bindingspecsfun_argmacrofu_Example_Unsafe_fooA@
 hs_bindgen_cf67e2fc00fd28d8 ::
-     A
+     RIP.FunPtr A
   -> IO ()
 hs_bindgen_cf67e2fc00fd28d8 =
   RIP.fromFFIType hs_bindgen_cf67e2fc00fd28d8_base
@@ -93,19 +93,19 @@ hs_bindgen_cf67e2fc00fd28d8 =
     __exported by:__ @binding-specs\/fun_arg\/macro\/function.h@
 -}
 fooA ::
-     A
+     RIP.FunPtr A
      -- ^ __C declaration:__ @x@
   -> IO ()
 fooA = hs_bindgen_cf67e2fc00fd28d8
 
 -- __unique:__ @test_bindingspecsfun_argmacrofu_Example_Unsafe_fooB@
 foreign import ccall unsafe "hs_bindgen_269a46f9680e33ed" hs_bindgen_269a46f9680e33ed_base ::
-     (RIP.Int32 -> IO RIP.Int32)
+     RIP.FunPtr RIP.Void
   -> IO ()
 
 -- __unique:__ @test_bindingspecsfun_argmacrofu_Example_Unsafe_fooB@
 hs_bindgen_269a46f9680e33ed ::
-     B
+     RIP.FunPtr B
   -> IO ()
 hs_bindgen_269a46f9680e33ed =
   RIP.fromFFIType hs_bindgen_269a46f9680e33ed_base
@@ -117,7 +117,7 @@ hs_bindgen_269a46f9680e33ed =
     __exported by:__ @binding-specs\/fun_arg\/macro\/function.h@
 -}
 fooB ::
-     B
+     RIP.FunPtr B
      -- ^ __C declaration:__ @x@
   -> IO ()
 fooB = hs_bindgen_269a46f9680e33ed