+++ hs-bindgen/src-internal/HsBindgen/Frontend/Pass/Final.hs
@@ -10,9 +10,9 @@ module HsBindgen.Frontend.Pass.Final (
 import GHC.Records
 
 import HsBindgen.Frontend.AST.Decl qualified as C
+import HsBindgen.Frontend.Pass.AdjustTypes.IsPass (AdjustTypes)
 import HsBindgen.Frontend.Pass.HandleMacros.IsPass
 import HsBindgen.Frontend.Pass.MangleNames.IsPass qualified as MangleNames
-import HsBindgen.Frontend.Pass.Select.IsPass
 
 {-------------------------------------------------------------------------------
   Definition
@@ -22,7 +22,7 @@ import HsBindgen.Frontend.Pass.Select.IsPass
 --
 -- Backend passes should refer to 'Final' instead of the actual name of the
 -- final pass, so that if we add more passes, the backend is unaffected.
-type Final = Select
+type Final = AdjustTypes
 
 {-------------------------------------------------------------------------------
   Annotations