+++ hs-bindgen/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * BREAKING: occurrences of the `CFieldType`/`CBitfieldType` type families in
   class instance heads are now replaced by their definition.
+* PATCH: fix a bug where `FunPtr`s were omitted for macro-defined newtypes.
 
 ## 0.1.0-alpha -- 2026-02-06